ports/esp32: Makefile produces static libs. main.c is modified in order to not include app_main() in the static libs. The same old behaviour is kept by moving the old main.c into new files named esp32task.h and esp32task.c
