+++ ports/esp32/esp32task.h
@@ -1,0 +1,7 @@
+#include "freertos/FreeRTOS.h"
+#include "freertos/task.h"
+
+void mp_task(void *);
+TaskHandle_t micropython_task(int);
+void nlr_jump_fail(void *);
+void mbedtls_debug_set_threshold(int);
